+++ database/migrations/2018_03_06_223058_create_team_users_table.php
@@ -4,23 +4,33 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateTeamUsersTable extends Migration
-{
+class CreateTeamUsersTable extends Migration {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
+    public function up() {
         Schema::create('team_users', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('team_id')->unsigned()->index();
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table
+                ->integer('team_id')
+                ->unsigned()
+                ->index();
+            $table
+                ->integer('user_id')
+                ->unsigned()
+                ->index();
             $table->timestamps();
 
-            $table->foreign('team_id')->references('id')->on('teams');
-            $table->foreign('user_id')->references('id')->on('users');
+            $table
+                ->foreign('team_id')
+                ->references('id')
+                ->on('teams');
+            $table
+                ->foreign('user_id')
+                ->references('id')
+                ->on('users');
         });
     }
 
@@ -29,8 +39,7 @@ class CreateTeamUsersTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::dropIfExists('team_users');
     }
 }
