+++ database/migrations/2018_04_27_120315_create_company_users_table.php
@@ -4,23 +4,33 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateCompanyUsersTable extends Migration
-{
+class CreateCompanyUsersTable extends Migration {
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up()
-    {
+    public function up() {
         Schema::create('company_users', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
-            $table->integer('company_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table
+                ->integer('user_id')
+                ->unsigned()
+                ->index();
+            $table
+                ->integer('company_id')
+                ->unsigned()
+                ->index();
             $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->foreign('company_id')->references('id')->on('companies');
+            $table
+                ->foreign('user_id')
+                ->references('id')
+                ->on('users');
+            $table
+                ->foreign('company_id')
+                ->references('id')
+                ->on('companies');
         });
     }
 
@@ -29,8 +39,7 @@ class CreateCompanyUsersTable extends Migration
      *
      * @return void
      */
-    public function down()
-    {
+    public function down() {
         Schema::dropIfExists('company_users');
     }
 }
